pyproject.toml
[project]
name = "extraction-review"
version = "0.1.0"
description = "Extracts data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-workflows>=2.13.0,<3.0.0",
    "python-dotenv>=1.1.0",
    "jsonref>=1.1.0",
    "click>=8.2.1,<8.3.0",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.0",
    "respx>=0.22.0,<1",
    "llama-cloud>=1.3.0",
    "json-schema-to-pydantic>=0.4.8",
    "pypdf>=6.9.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
    "typescript>=0.0.12",
    "ty>=0.0.2",
    "pytest>=8.4.1",
    "hatch>=1.14.1",
    "pytest-asyncio>=1.3.0",
    "playwright>=1.56.0,<2",
    "psutil>=7.1.3",
    "pytest-playwright>=0.7.2",
    "pytest-timeout>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.envs.default.scripts]
"format" = "ruff format ."
"format-check" = "ruff format --check ."
"lint" = "ruff check --fix ."
"lint-check" = ["ruff check ."]
typecheck = "ty check src"
test = "pytest"
"all-check" = ["format-check", "lint-check", "test"]
"all-fix" = ["format", "lint", "test"]

[tool.pytest.ini_options]
timeout = 120
timeout_method = "thread"

[tool.llamadeploy]
env_files = [".env"]
llama_cloud = true

[tool.llamadeploy.workflows]
process-file = "extraction_review.process_file:workflow"
metadata = "extraction_review.metadata_workflow:workflow"

[tool.llamadeploy.ui]
directory = "ui"

